+++ packages/pluggableWidgets/combobox-web/src/components/MultiSelection/MultiSelection.tsx
@@ -15,6 +15,7 @@ export function MultiSelection({
     tabIndex,
     a11yConfig,
     menuFooterContent,
+    ariaRequired,
     ...options
 }: SelectionBaseProps<MultiSelector>): ReactElement {
     const {
@@ -130,7 +131,8 @@ export function MultiSelection({
                                 }
                             },
                             disabled: selector.readOnly,
-                            readOnly: selector.options.filterType === "none"
+                            readOnly: selector.options.filterType === "none",
+                            "aria-required": ariaRequired
                         })}
                     />
                     <InputPlaceholder isEmpty={selectedItems.length <= 0}>{memoizedselectedCaptions}</InputPlaceholder>