+++ packages/pluggableWidgets/combobox-web/src/helpers/getSelector.ts
@@ -1,7 +1,8 @@
 import { ComboboxContainerProps } from "../../typings/ComboboxProps";
 import { AssociationMultiSelector } from "./Association/AssociationMultiSelector";
 import { AssociationSingleSelector } from "./Association/AssociationSingleSelector";
-import { DatabaseSingleSelector } from "./Database/DatabaseSingleSelector";
+import { DatabaseMultiSelectionSelector } from "./Database/DatabaseMultiSelectionSelector";
+import { DatabaseSingleSelectionSelector } from "./Database/DatabaseSingleSelectionSelector";
 import { EnumBooleanSingleSelector } from "./EnumBool/EnumBoolSingleSelector";
 import { StaticSingleSelector } from "./Static/StaticSingleSelector";
 import { Selector } from "./types";
@@ -18,7 +19,11 @@ export function getSelector(props: ComboboxContainerProps): Selector {
             throw new Error(`'optionsSourceType' of type '${props.optionsSourceType}' is not supported`);
         }
     } else if (props.source === "database") {
-        return new DatabaseSingleSelector();
+        if (props.optionsSourceDatabaseItemSelection?.type === "Multi") {
+            return new DatabaseMultiSelectionSelector();
+        } else {
+            return new DatabaseSingleSelectionSelector();
+        }
     } else if (props.source === "static") {
         return new StaticSingleSelector();
     }