+++ packages/pluggableWidgets/combobox-web/typings/ComboboxProps.d.ts
@@ -4,16 +4,16 @@
  * @author Mendix Widgets Framework Team
  */
 import { ComponentType, ReactNode } from "react";
-import { ActionValue, DynamicValue, EditableValue, ListValue, ListAttributeValue, ListExpressionValue, ListWidgetValue, ReferenceValue, ReferenceSetValue } from "mendix";
+import { ActionValue, DynamicValue, EditableValue, ListValue, ListAttributeValue, ListExpressionValue, ListWidgetValue, ReferenceValue, ReferenceSetValue, SelectionSingleValue, SelectionMultiValue } from "mendix";
 import { Big } from "big.js";
 
 export type SourceEnum = "context" | "database" | "static";
 
 export type OptionsSourceTypeEnum = "association" | "enumeration" | "boolean";
 
-export type OptionsSourceDatabaseCaptionTypeEnum = "attribute" | "expression";
-
 export type OptionsSourceAssociationCaptionTypeEnum = "attribute" | "expression";
+
+export type OptionsSourceDatabaseCaptionTypeEnum = "attribute" | "expression";
 
 export interface OptionsSourceStaticDataSourceType {
     staticDataSourceValue: DynamicValue<string | Big | boolean | Date>;
@@ -21,7 +21,7 @@ export interface OptionsSourceStaticDataSourceType {
     staticDataSourceCaption: DynamicValue<string>;
 }
 
-export type FilterTypeEnum = "contains" | "startsWith" | "none";
+export type FilterTypeEnum = "contains" | "containsExact" | "startsWith" | "none";
 
 export type OptionsSourceAssociationCustomContentTypeEnum = "yes" | "listItem" | "no";
 
@@ -52,17 +52,18 @@ export interface ComboboxContainerProps {
     attributeEnumeration: EditableValue<string>;
     attributeBoolean: EditableValue<boolean>;
     optionsSourceDatabaseDataSource?: ListValue;
-    optionsSourceDatabaseCaptionType: OptionsSourceDatabaseCaptionTypeEnum;
-    optionsSourceDatabaseCaptionAttribute?: ListAttributeValue<string>;
-    optionsSourceDatabaseCaptionExpression?: ListExpressionValue<string>;
+    optionsSourceDatabaseItemSelection?: SelectionSingleValue | SelectionMultiValue;
     optionsSourceDatabaseValueAttribute?: ListAttributeValue<string | Big>;
-    databaseAttributeString: EditableValue<string | Big>;
-    optionsSourceDatabaseDefaultValue: DynamicValue<string | Big>;
+    databaseAttributeString?: EditableValue<string | Big>;
+    optionsSourceDatabaseDefaultValue?: DynamicValue<string | Big>;
     attributeAssociation: ReferenceValue | ReferenceSetValue;
     optionsSourceAssociationDataSource?: ListValue;
     optionsSourceAssociationCaptionType: OptionsSourceAssociationCaptionTypeEnum;
+    optionsSourceDatabaseCaptionType: OptionsSourceDatabaseCaptionTypeEnum;
     optionsSourceAssociationCaptionAttribute?: ListAttributeValue<string>;
+    optionsSourceDatabaseCaptionAttribute?: ListAttributeValue<string>;
     optionsSourceAssociationCaptionExpression?: ListExpressionValue<string>;
+    optionsSourceDatabaseCaptionExpression?: ListExpressionValue<string>;
     staticAttribute: EditableValue<string | Big | boolean | Date>;
     optionsSourceStaticDataSource: OptionsSourceStaticDataSourceType[];
     emptyOptionText?: DynamicValue<string>;
@@ -101,17 +102,18 @@ export interface ComboboxPreviewProps {
     attributeEnumeration: string;
     attributeBoolean: string;
     optionsSourceDatabaseDataSource: {} | { caption: string } | { type: string } | null;
-    optionsSourceDatabaseCaptionType: OptionsSourceDatabaseCaptionTypeEnum;
-    optionsSourceDatabaseCaptionAttribute: string;
-    optionsSourceDatabaseCaptionExpression: string;
+    optionsSourceDatabaseItemSelection: "Single" | "Multi" | "None";
     optionsSourceDatabaseValueAttribute: string;
     databaseAttributeString: string;
     optionsSourceDatabaseDefaultValue: string;
     attributeAssociation: string;
     optionsSourceAssociationDataSource: {} | { caption: string } | { type: string } | null;
     optionsSourceAssociationCaptionType: OptionsSourceAssociationCaptionTypeEnum;
+    optionsSourceDatabaseCaptionType: OptionsSourceDatabaseCaptionTypeEnum;
     optionsSourceAssociationCaptionAttribute: string;
+    optionsSourceDatabaseCaptionAttribute: string;
     optionsSourceAssociationCaptionExpression: string;
+    optionsSourceDatabaseCaptionExpression: string;
     staticAttribute: string;
     optionsSourceStaticDataSource: OptionsSourceStaticDataSourcePreviewType[];
     emptyOptionText: string;