+++ packages/pluggableWidgets/datagrid-number-filter-web/e2e/DataGridNumberFilter.spec.js
@@ -1,4 +1,5 @@
 import { test, expect } from "@playwright/test";
+import AxeBuilder from "@axe-core/playwright";
 
 test.afterEach("Cleanup session", async ({ page }) => {
     // Because the test isolation that will open a new session for every test executed, and that exceeds Mendix's license limit of 5 sessions, so we need to force logout after each test.
@@ -54,22 +55,26 @@ test.describe("datagrid-number-filter-web", () => {
     });
 
     test.describe("a11y testing:", () => {
-        test.fixme("checks accessibility violations", async ({ page }) => {
+        test("checks accessibility violations", async ({ page }) => {
             await page.goto("/");
-            await page.initializeAccessibility();
-            await page.setAccessibilityOptions({
-                rules: [
-                    { id: "aria-required-children", reviewOnFail: true },
-                    { id: "label", reviewOnFail: true },
-                    { id: "aria-roles", reviewOnFail: true },
-                    { id: "button-name", reviewOnFail: true },
-                    { id: "duplicate-id-active", reviewOnFail: true },
-                    { id: "duplicate-id", reviewOnFail: true }
-                ]
-            });
+            await page.waitForLoadState("networkidle");
 
-            const snapshot = await page.accessibilitySnapshot({ root: ".mx-name-datagrid1" });
-            await expect(snapshot).toHaveNoAccessibilityIssuesMatching({ type: "wcag2a" });
+            const accessibilityScanResults = await new AxeBuilder({ page })
+                .withTags(["wcag21aa"])
+                .disableRules([
+                    "aria-required-children",
+                    "label",
+                    "aria-roles",
+                    "button-name",
+                    "duplicate-id-active",
+                    "duplicate-id",
+                    "aria-allowed-attr",
+                    "image-alt"
+                ])
+                .exclude(".mx-name-navigationTree3")
+                .analyze();
+
+            expect(accessibilityScanResults.violations).toEqual([]);
         });
     });
 });