+++ packages/pluggableWidgets/datagrid-web/e2e/DataGridSelection.spec.js
@@ -1,4 +1,5 @@
 import { test, expect } from "@playwright/test";
+import AxeBuilder from "@axe-core/playwright";
 
 test.afterEach("Cleanup session", async ({ page }) => {
     // Because the test isolation that will open a new session for every test executed, and that exceeds Mendix's license limit of 5 sessions, so we need to force logout after each test.
@@ -54,26 +55,26 @@ test.describe("datagrid-web selection", async () => {
         await expect(page).toHaveScreenshot(`datagridMultiSelectionRowClick.png`);
     });
 
-    test.fixme("checks accessibility violations", async ({ page }) => {
+    test("checks accessibility violations", async ({ page }) => {
         await page.goto("/p/multi-selection");
-        await page.initializeAccessibility();
-        await page.setAccessibilityOptions({
-            rules: [
-                { id: "aria-required-children", reviewOnFail: true },
-                { id: "label", reviewOnFail: true }
-            ]
-        });
+        await page.waitForLoadState("networkidle");
 
-        const multiSelectionCheckbox = page.locator(".mx-name-dgMultiSelectionCheckbox");
-        const report = await multiSelectionCheckbox.accessibilitySnapshot({
-            runOnly: {
-                type: "tag",
-                values: ["wcag2a"]
-            }
-        });
+        await page.locator(".mx-name-dgMultiSelectionCheckbox").waitFor();
+        const accessibilityScanResults = await new AxeBuilder({ page })
+            .withTags(["wcag21aa"])
+            .disableRules([
+                "aria-required-children",
+                "label",
+                "aria-roles",
+                "button-name",
+                "duplicate-id-active",
+                "duplicate-id",
+                "aria-allowed-attr"
+            ])
+            .include(".mx-name-dgMultiSelectionCheckbox")
+            .exclude(".mx-name-navigationTree3")
+            .analyze();
 
-        for (const violation of report.violations) {
-            console.log(`Violation: ${violation.description}`);
-        }
+        expect(accessibilityScanResults.violations).toEqual([]);
     });
 });