+++ packages/pluggableWidgets/datagrid-web/src/features/data-export/DSExportRequest.ts
@@ -172,7 +172,7 @@ export class DSExportRequest {
                 } else {
                     status = "available";
                 }
-                return status === "available";
+                return status !== "loading";
             });
         });